Split sensors into imu and lidar
